[kwin] Streamline Comment fields of KCMs
This patch applies a common language and type-setting to the
systemsettings modules in kwin.

Considerations:
- The comment field might repeat the name, or give more detail about the
  specific settings on this page, this makes sense with how
  systemsettings and kcmshell present it
- Mentioning the words settings, configure, options, etc. is avoided --
  it is clear from the context that these are settings and options.
- Title-case throughout in line with human interface guidelines, see
  http://techbase.kde.org/Projects/Usability/HIG/Capitalization
- The comment ends up being the title, so the
- tech slang is avoided as much as possible, but left in where really
  necessary
- I've left the Name field "mostly untouched", as that one is key for
  the user to find the right module in systemsettings' icon view and in
  the sidebars

The most prominent change is "Compositor" to "Graphics Compositor" (in
the Title field). This is more like a "suggestion", if you don't like
it, I'll take it out again. Reasoning: "Compositor" is a very technical
term, people who know graphics might recognize it. Others might not
understand that "composite" is about graphics -- hence the
specialization. Please enter the commit message for your changes. Lines starting

REVIEW:118338