[kcmkwin/kwincompositing] Drop the OpenGLErrorView from Effects
As we have the split between Effects and Compositor settings nowadays
it's incorrect to have an Compositor related option exposed in the
Effects settings view. Thus the OpenGLErrorView should be removed. As
a nice side-effect that simplifies the main qml file quite a bit by
turning the EffectView from Item into Rectangle.

If it shows that we need the reset GL check in the effects view it would
be better to use a KMessageWidget like in the compositing view.

BUG: 336342
REVIEW: 118790