Revert "set QSG_RENDER_LOOP as basic if not explicitly set"
This reverts commit c09d63d1928a27879487ef78e0adc6fc02b41c30.

The test results we got indicate that on NVIDIA the problems with the
threaded rendering loop do not exist and at the same time the rendering
result improves a lot. As Qt still disables the threaded rendering loop
for the problematic drivers, it's better to not force the basic rendering
loop.